Add a basic admin interface
With the admin interface (avaliable at /admin/model/YOUR_MODEL) you can view, add, and delete records for any model through a web browser.

The interface is clunky at the moment but there is support for entering text fields, viewing associations, and specifying whether a time field should be set to "now" or not.

Related changes with this commit

 * Updated version of ErlyDTL

 * boss_db:count/2 returns the number of matching BossRecords

 * New format for specifying a default actions in boss.config... this will break your app

  ** front_page tells which controller/action should be shown at /
  ** default_action tells which action should be the default for any controller (e.g. /forum)
  ** default_actions (note the "s") tells which action should be the default for specific controllers

 * Updated website docs

 * Updated "Hello world!" so you can play with it using the Admin interface