Permit integers, floats, binaries in BossRecords
Updated the tyrant driver to serialize/deserialize integers, floats,
binaries, and DateTimes, so invocations to :new are not restricted
to strings anymore. The "schema" is stored on a per-record basis.
Sorting will work as expected for all supported data types.

Since DateTimes are permitted in any attribute (not just those that
end in Time), I'll probably rip out the rules about attributes
ending in Time -- they're left in for now to preserve compatibility.