Boss Translator: easy i18n
Everything needed for basic i18n; see /admin/lang on a dev server to
check it out. The most important component is boss_translator, which is
a gen_server that translates messages. ErlyDTL now takes a translation
fun at render-time, and Boss passes in a fun which will call
boss_translator:lookup with the appropriate language. Language is
inferred from the Accept-Language header using reported q-values and
the percent of template strings for which a translation is available
in a particular language. (But the inferred language can be overriden
by setting a Content-Language header in the controller.) Messages are
stored as PO-files in Lang/.