Allow selection of test modules to run
The run_tests/1 function can now be called with a list of arguments, which first
element is the database adapter and subsequent elements are test
modules to run. Example:

  $ erl -noshell -pa ../ChicagoBoss/ebin \
        -run boss_web_test run_tests mock mail_test account_test

This allows user to select which select which tests they want to run. This can
be useful when the test suite is getting big.