Unicode support for HTTP POST data
NOTES
-----
Fixed and tested only with Misultin, not with MochiWeb.

PROBLEM DESCRIPTION
-------------------
Once we submit a string like "koníček" the browser
urlencodes it as "kon%C3%AD%C4%8Dek" what is correct.

By default, before applying this changeset, ChicagoBoss
received string like
  [107,111,110,   195,173,   196,141,   101,107],
that may be in some context correct too. The problem
is that Unicode is family of multibyte charsets and
server does not know that 195,173 and 196,141 are
not four, but two characters.

Once we type "koníček" down to the Erlang shell we
will get string like
  [107,111,110,   237,   269,   101,107]
that is correct.

SOLUTION
--------
After applying this changeset, Misultin will respect
that we work in UTF-8 mode and will translate
"kon%C3%AD%C4%8Dek" correctly to [107,111,110,237,269,101,107].

CAVEATS
-------
The only problem is that if we need to return such
a Unicode string back to client's socket, we have to
wrap our string in unicode:characters_to_binary ()
function which is bothering and should be automatized
somehow (in ErlyDTL variable expansion logic??;
it will be much more clear to do it within Misultin
Response).