Add include_dirs config option
Adds a configuration option to allow -including of files from
directories other than the default directory. This is useful (indeed,
necessary) when using ChicagoBoss as part of a larger project with, say,
types defined externally that you want to pull in. To use, add a line
like this to boss.config in the 'boss' section:

	{include_dirs, ["../app/src/include", "../some_other_app/include"]}