Semantic versioning support;
For those who use `rebar_vsn_plugin` for semantic versioning,
this is a patch for ChicagoBoss that integrates with our
fork of the `rebar_vsn_plugin`:
  https://github.com/ztmr/rebar_vsn_plugin
(The pull request back to the erlware is pending at the moment.)

WARNING: do not put `rebar_vsn_plugin` into your `rebar_plugins`
in `rebar.config` in the case of any CB applications -- it may
override the standard CB's plugin and you will end up with
broken `.app` file what breaks the app at all (at least missing
list of controllers and so on).

In all the non-CB apps, you can happily use `rebar_vsn_plugin`
as usual.