boss_session: mnesia initialization improved;
When the `session_adapter` is set to `mnesia`, the Mnesia
schema was not created automatically. This is to fix it
by adding the `ensure_schema` function.

Note that it has not been tested in distributed environment
yet so we expect the `ensure_schema` may become a bit more
complex in the future.

Additionaly, the timeout for `wait_for_tables` was
increased from 20ms to 1000ms because 20 was not enough
at least in our testing environment.