Pass a new _req template variable
Provide all the templates with a new `_req` variable
that holds the original web request so that template
can use query and post params directly.

There are situations where it is much more straightforward
than to pass the specific parameters via controller
all the time again and again, across the whole application.