Fix module_older_than when file not exists. Also..
Also remove unnecessary calendar:datetime_to_gregorian_seconds
conversion. Comparing datetime tuples should be sufficient, and also
automatically handles the case where a file is not found, since
filelib:last_modified returns `0` if the file is not found.

Also add a a few try-catch blocks:

1) To boss_load:compile_view_dir_erlydtl, where erlydtl:compile_dir was
crashing due to directory-not-found error.  This may not need to be
permanent, since it's contingent on a fix in erlydtl, but without it
right now, CB doesn't work.

2) To boss_web_controller_handle_request:handle_request/3, where a crash
was silently failing rather than producing an error message in any log.
It will now generate an appropriate lager error message so it can
be debugged.