boss_mail_driver_smtp: tune gen_smtp options for relays
It appears that some relays like Gmail have issues when MX lookups
are performed. This issue has been longstanding in general, even
after RFCs 973 and 974 standardized the record. As such, set
no_mx_lookups to true and fall back on A records.

In addition, be more lenient with retry attempts, from only 1
to 3. This may help with boss_mail timeouts to SMTP relays.